Refactor activity change tests into per-platform tests
